Added skeleton for command populatedb
In the requirements of the project, it was needed to present
some form of database so the proffessor could use the application.
But it's not considered good practices to upload a testing database,
as it can be an origin of merge conflicts and some other troublesome
problems. Instead, the best approach is to make an script to upload
some database. Luckly, following [this approach of making a command](https://docs.djangoproject.com/en/3.0/howto/custom-management-commands/)
can be useful. The idea behind is to make a command *populatedb*
as when executing `python3 manage.py populatedb` creates all
the instances necessary of the project. It was also created a
skeleton test for the command (it does not contain anything).