Added tests.
Copy the skeleton of this tests. Problems I found:
- When using LoginRequiredMixin the next option
makes the browser return to the page you were visiting (cool!),
but I forgot this feature in all tests, like in behave and normal
unit tests.
- context.get_url it's more intelligent than I thought. It uses
reverse if it is usable. This make a step in simplifying the tests
- Behave is slow, so the first step to make sure it works is test,
then open it in your browser, then behave.