Added Create band. Made some blank detail band
Create band was easy. The two methods overrided are
for storing the user who created the object, and to
get the success url, as it needed self.object (the
object created) to make sense. Added also the method
get_absolute_url at band, as it's used in splinter, behave,
and tests.