Changed tags to names.
Now, tags are unique, so when you put more than one weird shit
happens at splinter. In order to use behave with more than one
tag, "name" value is more cool. An other problem was checking the
text, so I changed to use h2 instead of the div#card as it was
better. Also added a bunch of frontend stuff so it looks cool.